Fix Shotgun

+++ Assets/Scripts/ItemScripts/Shotgun.cs
@@ -6,6 +6,6 @@ using UnityEngine;
 public class Shotgun : Gun
 {
     public int numberOfBullets;
-    [Range(0, 90)]
-    public int spreadAngle;
+    [Range(0f, 90f)]
+    public float spreadAngle;
 }
